feat: OmokDao 클래스 생성 및 게임 상태를 저장하고 불러오는 로직 분리
