+++ app/src/main/java/woowacourse/omok/OmokGameManager.kt
@@ -1,6 +1,6 @@
 package woowacourse.omok
 
-import data.DbHelper
+import data.OmokDao
 import domain.domain.Point
 import domain.domain.state.Foul
 import domain.domain.state.Playing
@@ -10,14 +10,14 @@ class OmokGameManager {
     fun updateState(
         state: State,
         point: Point,
-        dbHelper: DbHelper,
+        omokDao: OmokDao,
     ): State {
         if (state !is Playing) return state
 
         val newState = state.place(point, 15) { _, _ -> }
 
         if (newState !is Foul) {
-            dbHelper.saveGameState(newState)
+            omokDao.saveGameState(newState)
         }
 
         return newState
